新增搜尋商品功能，修正一些小錯誤、前端資料夾分類

+++ src/main/java/com/guohenry/springbootmall2/model/Product.java
@@ -6,7 +6,7 @@ public class Product implements Serializable {
     private int id;
     private String name;
     private String description;
-    private double price;
+    private Integer price;
     private String imageUrl;
 
     public int getId() {
@@ -33,11 +33,11 @@ public class Product implements Serializable {
         this.description = description;
     }
 
-    public double getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
